chore(team): split run admin manager tests
